Moved Frame into a file and added doc
